Fixed prek error and renamed checkmark_checkbox
